Clean up the room read/write code a bit.
Probably breaks everything, but it's 2AM and fundies is impatient, so fuck all.
Who needs code that works when you can have code that's healthy.
